+++ lib/kronix.dart
@@ -23,3 +23,5 @@ export 'database/middleware.dart';
 export 'database/migration.dart';
 export 'database/migration_runner.dart';
 export 'core/websocket.dart';
+export 'http/file.dart';
+export 'filesystem/storage.dart';